Use network_cmds' Header And Source Fild (conn_lib)
It's interesting that Apple doesn't have mptcp_client use libnetwork's implementation instead. Maybe it is an indication that mptcp_client requires the implementation to be different from what libnetwork provides.

Regardless, building the conn_lib.c in libnetwork seems to work, so I'll stick with this approach for now.

This is required to build nc (netcat).